Renaming user-remove-modal to user-remove-confirmation-modal
