+++ kolibri/plugins/facility_management/assets/src/views/class-edit-page/user-remove-confirmation-modal.vue
